POC: Enable Proton for XPU; basis, not ready for user use (#2635)
Closes #1145

On LTS proton doesn't work: `zelTracerCreate` return `2013265921`
(ZE_RESULT_ERROR_UNINITIALIZED) code. Looks like relates to
https://github.com/intel/intel-xpu-backend-for-triton/issues/1953.

**4/2/25 status update:**

There is still no necessary interface in PTI, but work on it is in
progress and it seems to me there is little left (thanks @jfedorov!). I
believe that at least the main points of the implementation can be
reviewed and maybe even merged. The main attention should be paid to
dependencies, which include sycl, level_zero (including an additional
interface for calling callbacks, which is used temporarily) and PTI. I
added PTI headers directly to the Triton sources, since they will be
small in size and are used only for building Proton, I believe this will
not interfere with the main code. I tried to add dependencies so as not
to break the build if they are missing, also some paths are hardcoded,
because the main use case for now is just launching UTs in CI. There is
a lot of debug code left in the code, which outputs additional
information to stdout, I would like to leave it for now. I plan to clean
this code as soon as PTI implements the necessary functionality (the
integration of which will highlight many problems, I'm sure, since I
can't emulate the behavior of the PTI completely).

**12/12/24 status update:**

The most important problem is the lack of an interface for registering
callback functions that would be called before and after kernel
execution. Cupti and Roctracer backend of Proton use this interface to
register (via `callbackData->correlationId`) ​​kernel calls to build a
call tree and, upon profiler shutdown, to check that all created records
for running kernels were written to Proton storage structures from the
backend profiler storage system. The current workaround uses the
interface `<level_zero/layers/zel_tracing_api.h>`, which provides
functions for registering user functions via a pair of
`prologue_callbacks/epilogue_callbacks` for various events. However, the
problem of obtaining the record identifier in these callbacks that PTI
will create for registered kernels **has not been solved yet** (a
manually selected identifier is currently used for tests).

How it works in Cupti? (simplified)
```c++
void CuptiProfiler::CuptiProfilerPimpl::callbackFn(void *userData,
                                                   CUpti_CallbackDomain domain,
                                                   CUpti_CallbackId cbId,
                                                   const void *cbData) {
// ...
  if (callbackData->callbackSite == CUPTI_API_ENTER) {
      // scope registration
      auto scopeId = threadState.record();
      threadState.enterOp(scopeId);
      // ...
      // linking the internal profiler data ID to the external one that proton uses. 
      profiler.correlation.correlate(callbackData->correlationId, numInstances);
  } else if (callbackData->callbackSite == CUPTI_API_EXIT) {
     // ...
     // scope exit
     threadState.exitOp();
     // the submitted record should be taken into account when flushing data
     profiler.correlation.submit(callbackData->correlationId);
  }
// ...
}

void CuptiProfiler::CuptiProfilerPimpl::doStart() {
  // subscriber is `CUpti_SubscriberHandle`
  cupti::subscribe<true>(&subscriber, callbackFn, nullptr);            // not exists in PTI
  cupti::activityEnable<true>(CUPTI_ACTIVITY_KIND_CONCURRENT_KERNEL);  // exists in PTI
  cupti::activityRegisterCallbacks<true>(allocBuffer, completeBuffer); // exists in PTI
  // `setGraphCallbacks` do something like:
  //   CALLBACK_ENABLE(CUPTI_CBID_RESOURCE_GRAPHNODE_CREATED);
  //   CALLBACK_ENABLE(CUPTI_CBID_RESOURCE_GRAPHNODE_CLONED);
  //   ...
  setGraphCallbacks(subscriber, /*enable=*/true);                      // not exists in PTI
  // `setRuntimeCallbacks` do something like:
  //   CALLBACK_ENABLE(CUPTI_RUNTIME_TRACE_CBID_cudaLaunch_v3020);
  //   CALLBACK_ENABLE(CUPTI_RUNTIME_TRACE_CBID_cudaLaunchKernel_v7000);
  //   ...
  setRuntimeCallbacks(subscriber, /*enable=*/true);                    // not exists in PTI
  // `setDriverCallbacks` do something like:
  //   CALLBACK_ENABLE(CUPTI_DRIVER_TRACE_CBID_cuLaunch);
  //   CALLBACK_ENABLE(CUPTI_DRIVER_TRACE_CBID_cuLaunchGrid);
  //   ...
  setDriverCallbacks(subscriber, /*enable=*/true);                     // not exists in PTI
}
``` 

**Needs to be done:**
* Think of what to do with getting the PTI profiler record identifiers
in level zero callbacks or how to find a way without them.
* ​​~Synchronize the device to make a correct data flush.~ (done via
`sycl_queue.wait()`)
* What to do with the concept of cuda graph kernels?
* Obtain the device architecture.
* Enable unit tests.
  * [x] test_api.py
  * [x] test_lib.py
  * [x] test_profile.py (partially)
  * [x] test_viewer.py
* Enable tutorials.
* Final code cleanup.



https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/14202903723

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>