[Intel] Rework load-store redundant data masking  (#3896)
Upstream updated the masking logic when generating vectorized
loads/stores to use linear layout free variable masks:
https://github.com/triton-lang/triton/pull/5432/files#diff-7147172e5de66b21e7447a220f435a703df5302dd91a41a5be9683e5396652c5

The free variable mask essentially finds inputs to the linear layout
that do not change the output - these inputs are duplicates and should
be masked when doing loads or stores. This PR applies that work to the
Intel backend. Previously, we were ignoring duplicates in the registers
which could result in incorrect results if the tensor size of the
load/store operand was smaller than the DPAS tile size in a particular
dimension. A nice side effect here is we generate fewer stores - from 32
`<1 x i16>` store instructions in the LLVM IR previously to 16
instructions now for the B operand in the float16 128x8 test case.

Close #3841