Fix the total number of tests for LTS (#3994)
Previously for LTS we didn't run the instrumentation test, because
pytest does not work when all tests are deselected. Now, with the new
pytest-skip plugin, the test is skipped, and it is safe to remove the
old code for LTS. This should fix the long living issue with the total
number of tests for LTS is less than for rolling.