Calculate block load tile layout dim size from bases (#3971)
Previously we attempted to compute the total size for the 2D block loads
according to the tensor layout using input parameters like tensor shape
and warp shape. This can be error prone, since the tensor shape is
manipulated according to the warp distribution. A cleaner solution is to
modify the dimension sizes according to the bases. By loading the
dimensions from the tile layout in addition to the bases, we can modify
the dimension sizes using the same metrics used to construct the bases.
This appears to be giving correct results using the `test_block_load`
tests. I did not use this approach initially because I was concerned
about the load tile being too "big", but because we incorporate strides
in the loads now this approach should faithfully represent the total
dimensionality of the loaded data.