Fix regression issue in flex decoding.  (#3999)
The `tt.store` operation with BlockPointer fallbacks to scatter store if
the BLOCK shape or the value layout was not supported by the 2D BLOCK
IO.
The lowering code would transform the BlockPointer to the pointers and
masks.

The scatter store should apply the `and` to the `maskElems` if the
`llMask` doesn't exsits.

Signed-off-by: Lu,Chengjun <chengjun.lu@intel.com>