Enable `SPV_INTEL_fp_fast_math_mode` (#4058)
Triton generates fast math flags like `contract`,
`SPV_INTEL_fp_fast_math_mode` SPIRV extension is required to have it
translated correctly, or else it would be dropped.

A failure was encountered with LTS driver with the addition of
`SPV_INTEL_fp_fast_math_mode` SPIRV extension, as the SPIRV translator
increases the SPIRV version to v1.6. An issue
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/issues/3137 is
created to resolve this problem, in the meantime, a workaround is added
in SPIRV translator.

Benchmark CI:
https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/14745401109
(No performance degradation.)

---------

Signed-off-by: Whitney Tsang <whitney.tsang@intel.com>