[LoadStoreOpToLLVM] Calculate pitch using stride from `AxisInfo` (#4097)
When performing lowering of `PrefetchOp` and `LoadOp` of pointers with
tensor of pointers type, the pointer difference of the first two
elements was used to determine the pitch,
`max(shuffle(ptrtoint(extractvalue ptrs[1]) - ptrtoint(extractvalue
ptrs[0]), 0), baseWidth)`.
This PR changes the lowering to use constant stride determined by
`AxisInfo` as the pitch when possible, instead of calculating
dynamically.

Benchmarks CI:
https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/14847658390
(no regressions)

---------

Signed-off-by: Whitney Tsang <whitney.tsang@intel.com>