Add attribute to control 16-bit atomics lowering. (#4149)
Currently, we always emulate 16-bit atomic operations. The new attribute
allows targets to disable the emulation.

With the growth of the number of feature attributes, I modified how we
pass them to the annotate module pass. Passing all features in as a
structure simplifies adding a new attribute, allows usage of default
values, and decreases merge conflicts.

Signed-off-by: Ilya Enkovich <ilya.enkovich@intel.com>