Use the actual runtime dependencies for PyTorch (#4218)
Updating "Test with pip" workflow that verifies that the latest nightly
wheels of PyTorch built in this repository works with Triton built from
main. Instead of DLE, the workflow uses runtime dependencies installed
from pypi. This changes ensures that these dependencies are actual
(fetched for pinned PyTorch commit) and not hardcoded in the workflow.

Signed-off-by: Pavel Chekin <pavel.chekin@intel.com>