Re-use configuration information from CUTLASS
Instead of creating our own mapping from problem shape to CUTLASS GEMM
configuration, re-use existing information in CUTLASS.

This adds a small tool that can parse CUTLASS' benchmark configuration
files and generate a C++ header with the problem shape to configuration
mapping.

The generated header is included in the CUTLASS kernel benchmark to
dispatch to the best known configuration for each problem shape.

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>