Fix warp num lookup segfault for empty modules (#4322)
There's a subtle problem with the pattern application for the conversion
to llvm pass I noticed recently. The symptom is that when I have two
modules that are lowered to llvm correctly separately, and I put them
into a single test case, the matcher fails to find legalization
patterns. While trying to create a minimal reproducer, I came across
this issue that if I just add an empty module, our conversion fails with
a segfault. This small fix also helps with the initial problem I had
(although I'm not exactly sure why just yet).