[TritonGENToLLVM] Adjust base pointer in 2D block IO (#4335)
2D block IO instructions requires base pointer to be 64 bytes aligned.
When base pointer is not 64 bytes aligned, we need to make it 64 bytes
aligned and adjust base width and offset x accordingly to make it
accesses the desire memory. To make a non 64 bytes align pointer to be
64 bytes align, we clear the lower 6 bits.

We didn't clear the lower 6 bits before, because it is ignored on PVC,
and we can save on the number of instructions, but it is not ignored on
BMG, hence the incorrect behavior. We should not rely on a undocumented
behavior, and explicitly perform the action in Triton.

Fixes #4304

`Build and test B580, Linux` CI:
https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/15290645209

`Triton benchmarks BMG` CI:
https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/15308884810
GEMM tensor of pointer benchmark is fixed, no functional regression on
other benchmarks.

`Triton benchmarks PVC` CI:
https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/15290677510
Potential improvement on flex attention, no geomean regression on other
benchmarks.
![Screenshot 2025-05-28
100851](https://github.com/user-attachments/assets/4e5ab000-015e-461b-8f12-6ad44600de30)

---------

Signed-off-by: Whitney Tsang <whitney.tsang@intel.com>