Add subgroup 2d block load lit tests (#4328)
Adds lit tests for various subgroup 2d block IO ops. I created the tests
using the following strategy:

1. start with a simple layout, single warp, and vary the repCluster size
to change the size of the load / number of loads.
2. add the layouts from the AxB and AxBT gemm benchmark (AT does not use
2d block IO).
3. add the layouts from flex attention with custom masks 

I believe this gives us a solid baseline, though I am planning to use
this to debug another branch which is failing in the benchmarks but
passing existing UTs so I may add a small number of additional cases but
did not want to be duplicative initially.

Close #4253