Implement `fresh_triton_cache` fixture through `compilation.always_compile` (#4363)
The main idea is that for tests where we need the cache to be generated
each time, we can set this behavior via a special environment variable
(`TRITON_ALWAYS_COMPILE` or corresponding knobs), instead of creating a
new temporary folder and deleting it each time, which is problematic on
Windows. It seems like this solution is more lightweight, maybe it will
be possible to upstream it.

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>