[TritonGENToLLVM] Do not adjust zero base width (#4329)
When base shape to a 2d block io builtins is zero, the HW doesn't
perform an out-of-bound check in 2D block IO instructions (because it
doesn't know the actual size of the memory accessed).
Because the ptr operand may not be 64-bytes aligned, we compensate by
adjusting the base width and the offset X passed to the builtin.
However, this causes the base width to no longer be zero, and so the HW
assumes that the updated base width is the full memory width, and
consequently it ends up loading nothing.
Note: zero base width is used in our raise block pointer pass.

---------

Signed-off-by: Whitney Tsang <whitney.tsang@intel.com>