Use Python 3.10 in CI (#4428)
We have new test time dependencies that do not exist for Python 3.9
(#4424).

Note 1: we will continue building the wheels for 3.9 and will test with
Python 3.9 with ["Build and test
Python"](https://github.com/intel/intel-xpu-backend-for-triton/actions/workflows/build-test-python.yml)
(going to skip tests that reuire the new dependency, will do in a
separate PR).

Note 2: we need to update branch protection rules before merging this
PR.

Signed-off-by: Pavel Chekin <pavel.chekin@intel.com>