Introduce Subgroup 2D Block Encoding (#4193)
Add a new layout to describe the tensor layout with respect to the GPU
compute hierarchy (register, lane, warp, block). This PR introduces the
layout and adds its definition and basic functions to the Triton Intel
GPU Dialect. The conversion to Linear Layout function has been added and
unit tested through an Intel specific `LinearLayoutConversionsTest`. The
layouts are unpacked - each register is assumed to be the size of the
tensor type. However, the layout generation follows the convention
described in
https://github.khronos.org/SPIRV-Registry/extensions/INTEL/[SPV_INTEL_2d_block_io](https://github.khronos.org/SPIRV-Registry/extensions/INTEL/SPV_INTEL_2d_block_io.html).html.
While there may be some bugs, the goal is for any valid operation
described in the SPIRV extension to be represented correctly with this
layout.

Currently the layout is unused other than for linear layout conversion
testing purposes. I plan to leave this PR in draft until I have replaced
the `block_io` attribute on the load ops with this layout - and then I
plan to replace the linear layout code I added to
`LoadStoreOpToLLVM.cpp`. That second task might prove challenging since
I think the DPAS layouts do sometimes incorporate register packing
schemes into the layout - but looking at the upstream layouts for NVIDIA
and AMD MMA, specific packing is an implementation detail and not
represented as part of the high-level layout encoding.

cc #4192