Fix sub-group size used for shuffles. (#4490)
This patch fixes the subgroup size used for shuffles in conversion
layout lowering. This parameter should always match the number of
threads set for the module. Otherwise, the shuffle op lowering to SPIR
dialect would fail.

The patch also adds a special guarding pattern to check that our layout
conversion pattern succeeded in those cases where it's critical. Membar
analysis works on the assumption that these patterns would succeed, and
therefore, some of the barriers are optimized out. If we fail in our
lowering and use the generic pattern, then missing barriers might cause
incorrect behavior. This guard pattern helps to reveal that we have a
problem in a few variants of the `test_reduce_layout` and will allow us
to prevent regressions here in the future.

Fixes #4481

Signed-off-by: Ilya Enkovich <ilya.enkovich@intel.com>