Move tile size computation for subgroup 2d block io encoding  (#4461)
Moves the tile size computation from LLVM lowering to a static method on
the Subgroup 2D block encoding layout. This allows us to create the
layout with the desired tile sizes at a higher level in the pass
hierarchy. Moving this functionality now allows us to test using the
existing layouts and lowering, ensuring no regressions.

There is some cleanup that could be done but I opted for generic objects
for now (e.g. `SmallVector`) for flexibility.