Add regression test for `df19f1d` using problematic kernel from one of crashed models in #4410 (#4529)
When testing e2e models on PyTorch side, [an issue was
found](https://github.com/intel/intel-xpu-backend-for-triton/issues/4410)
that caused a crash. This issue was not detected by our testing, so the
idea arose to isolate the problematic kernel from one model and add it
to our test suite to be able to detect it faster and not allow
regressions.

The crash:
```bash
    # L0 build module failed. Log: IGC: Internal Compiler Error: Segmentation violation
    # Error during Intel loadBinary: Triton Error [ZE]: 0x70000004
    # RuntimeError: Triton Error [ZE]: 0x70000004
```

The commit that caused this issue:
https://github.com/intel/intel-xpu-backend-for-triton/commit/df19f1d314223163fb0a1620ecce8e081e2304d7

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>