Fix crash when processing getMakeTensorPtr op through while loop (#4567)
Properly handle while loop in `getMakeTensorPtr`. 

This is common code, but given upstream is deprecating the feature I
don't think they want to take this change. The test for this is in #4463
but I think it is better to keep the commit logically separate and not
have it squashed and merged with #4463.