[LoadStoreOpToLLVM] Fix load with `base height == 1` (#4602)
When `strides[0]` is 0, we only want to load the first row, so we set
the base height to be 1. (<= done in another PR)
When base height is less than tile height and base height is 1, only the
first row contain valid data.
To ensure the entire tile is filled with valid data, we must replicate
the first row throughout the tile.

---------

Signed-off-by: Whitney Tsang <whitney.tsang@intel.com>