Fix implementation of `test_tensor_descriptor_load_nd` (#4713)
The `test_tensor_descriptor_load_nd` test case contains the following
code:
```
  inp = to_triton(numpy_random(alloc_shape, dtype_str), device=device, dst_type=dtype_str)
  inp.data = inp.data[..., :INNER_BLOCK - 3]
```

The `inp.data` is updated (shallow update) to a tensor of 29 elements
(when INNER_BLOCK == 32), however `inp.shape` becomes stale as it was
created during construction of the `TensorWrapper` class during the call
to `to_triton`.

A better implementation is to slice the numpy array before translating
it to a triton tensor.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>