[BENCHMARK] Reuse CUTLASS's gemm configuration file (#4720)
# Description

This PR introduces a new mechanism for fetching GEMM configurations.

Instead of hardcoding the `(shape → config)` mapping, the
`config-tool.py` script now parses a configuration file and generates
the `gemm_config` structure dynamically.

The configuration file consists of a list of GEMM kernel invocations
with the corresponding `GemmConfig`. These will be extracted and used to
invoke the kernel with the appropriate configuration.

## Note

Currently, the configuration file is located in
`benchmarks/cutlass_kernel/gemm`. In the future, this should be updated
to fetch the file directly from the CUTLASS repository:
https://github.com/intel/cutlass-sycl
This change will be made once the CUTLASS repo includes a unified file
containing the optimal configurations for all shapes used in the Triton
benchmark.

---------

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>
Signed-off-by: Jefferson Le Quellec <jefferson.lequellec@codeplay.com>
Co-authored-by: Lukas Sommer <lukas.sommer@codeplay.com>