Avoid pass failure for descriptor load/store operations using a tensor descr. allocated on the host (#4812)
When a tensor descriptor is allocated on the host and passed to a triton
kernel as a argument, the pass that transforms descriptor load/store
operations into equivalent block pointer operations currently fails.
This PR rectifies the situation and allow the pass to complete so that
subsequently those operations can be lowered to use the 'unwrapped'
descriptor argument passed to the kernel.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>