Limit the minimum repeatCount value to make sure the number elements of DPAS operand A is enough to be shared to each logical lane. (#4787)
The number of elements of the matrix for dot operands A maybe less than
the sub-group-size.
E.G for the case: repCount = 1, systolic depth = 8, ops per channel = 1
and threads_per_warp = 16.

It is not supported by the IGC scalar backend. We have to limit the
minimal number of repCount=2 for that case with paddings.

Signed-off-by: Lu,Chengjun <chengjun.lu@intel.com>