Fixed FP to FP8 downcast (#4748)
Fixes #4630

This implementation is based on an arithmetic approach (see below) and
provides an RTNE conversion from F32/16 to F8 types. Despite it is using
more expensive fp binary operations, I don't see any performance
degradation (you can find a simple bench commented out in the test). The
implementation is very simple and template-based. It replaces 4 existing
functions that are more complicated and adds a direct conversion from
F32 to F8 without intermediate F16.

The implementation is based on the following approach. To convert a
float from src to dst, we need to find such an exponent and mantissa
that src is equal or close to dst. For the normal numbers (exp != 0) it
means:
```python
2^(srcExp - SrcBias) * (1 + srcMan/2^SrcMBits) = 2^(dstExp - DstBias) * (1 + dstMan/2^DstMBits)
```
and the following for subnormals (dstExp == 0):
```python
src = 2^(1 - DstBias) * (dstMan/2^DstMBits)
```
The exponent is calculated as:
```python
dstExp = max(0, srcExp - SrcBias + DstBias)
```
Simplifying the first formula, we can get the following:
```python
dstMan = srcMan * 2^(DstMBits - SrcMBits)
```
If `SrcBias == DstBias` (in the case of FP16 to F8E5 conversion), this
formula also works for subnormals. In the general case, we can use the
following formula for subnormals:
```python
dstMan = src * 2^(DstMBits + DstBias - 1)
```
Thus, to get the mantissa, we are using a simple multiplication by a
constant and rounding the result to the nearest int. In case of the
mantissa overflow, we need to reset it to zero and increment the
exponent. It could be done in the following way:
```python
dst = max((dstExp << DstMBits) + dstMan, DST_MAX)
```
It gives us the required dst value without the sign bit.