[CI][E2E accuracy] Add report aggregation and comparison with reference for E2E accuracy (#4939)
Implements
https://github.com/intel/intel-xpu-backend-for-triton/issues/4455

What is done:
1. Generation of concatenated report file (csv) that contains all rows
from separate csv files for each configuration
2. Generation of aggregated report file (csv) that contains ~45 rows for
all (suite, mode, dtype) configuration with pass rate and failed models.
Also, same report is printed as an output
3. Comparison against reference results from
https://github.com/intel/torch-xpu-ops/blob/main/.github/ci_expected_accuracy/check_expected.py

Comparison against reference is not called if E2E is called with
specific model or with a subset of models.