Implement `make_opt_flags` function for XPU and enable tests in `test_matmul.py` (#5051)
**Note for reviewers:** I've left only the most basic heuristics. If you
have improvements in mind that will definitely work better without
testing, i.e. were already known, then we can make edits directly to
this pull request. If you have improvements that need to be tested -
that's also good, please also write, but I'd prefer to implement the
basic version as quickly as possible and tune it in separate PRs if
possible.

Pass rate: 84.11% -> 89.04%

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>