[CI][E2E-accuracy] Fix bug with aggregation of 1 result (#5169)
Closes https://github.com/intel/intel-xpu-backend-for-triton/issues/5158

The issue appeared because [download
artifact](https://github.com/actions/download-artifact?tab=readme-ov-file#v5---whats-new)
will change it's behavior if only one artefact is available. In this
case the we will loose the artifact name and extract it's content into
selected folder. So it's `separate-reports/contentA` instead of
`separate-reports/artifactA/contentA;separate-reports/artifactB/contentB`.

The change is that we now always merge artifacts and lose names like
`artifactA`. So we have to change result parsing as well.