[Skiplists] Reduce the length of the skiplists with regex  (#5237)
Addressing #5108, this PR introduces skiplist reductions with the use of
new features in the pytest-skip plugin.

It was possible to significantly reduce the lengths for most of the
skiplists, however there are still certain groups of test cases that
remained untouched due to regex pattern creation complexity.
Technically, all lists can be matched with a single pattern, however
this comes at a trade-off of the pattern readibility.
The number of skips for the tests affected, was manually checked whether
it alignes with the main branch without the regex patterns and no
additional tests are skipped. Below is the summary:
<img width="600" height="600" alt="image"
src="https://github.com/user-attachments/assets/2c892188-ed8d-4ad8-8c19-d0af0d1682ee"
/>