Integrate PTI callback interface and build it from sources (#5289)
Closes #3852 

The first step to obtaining profiling data in more complex cases where
the previous workaround did not work.

The assembly of PTI wheels will be organized separately, since PTI is
not yet ready for this (there is no setup.py/pyproject.toml). FYI
@jfedorov

The next immediate step is to improve the pass rate and clean up old
code.

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>