A tracking utility for gathering the compile and/or runtime time, size, profiling and other statistics  (#4777)
To enable the tracking, set the environment variable
``TRITON_TRACK_DUMP`` to either ``1``, ``true``, ``yes``, ``on``, ``y``
or a path to a directory where the tracking reports will be dumped.
To add the profiling statistics to the reports, set the
``TRITON_TRACK_PROFILE`` environment variable.
To track the kernel launches, set the ``TRITON_TRACK_RUN`` environment
variable.

Link #4716