Create a Fuser utility designed to fuse operations in the same DefUseChain (#5332)
The Triton XPU compiler has implemented a couple of transformations
patterns to fuse "adjacent" operations (e.g. fuse a transpose with a
load, fuse a reshape with a load). Both of these transformation leverage
the `DefUseChain` infrastructure already built. This PR consolidate
common functionality into a generic Fuser class so that derived classes
can specialize the fusing transformation and reuse the base class.

---------

Signed-off-by: Ettore Tiotto <ettore.tiotto@intel.com>