Implement benchmarking using Proton (#5385)
This pull request adds a new profiling mode: `PROTON_PROFILER`, as well
as an optional run of the `fused_softmax` benchmark with this mode as a
sanity check (collected data is not recorded, but only output to
stdout). After further testing, it can be enabled for all benchmarks and
made the default benchmarking mode.

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>