Remove unnecessary masks (#5384)
This PR extends the `RemoveMask` pass in order to consider mask on load
and select operations that evaluate to true (or false) in the entire
loop iteration space. This masked loads to be transformed into unmasked
ones, and the mask condition may become dead if not used by other
operations (therefore it may contribute to reduction of arithmetic
complexity).

---------

Signed-off-by: Ettore Tiotto <ettore.tiotto@intel.com>