Added new JGS workflows to add them to workflow menu (#694)
New JGS CI needs to be added to main-interim so that it appears in the
workflow menu. Unfortunately, it means that it will run on cron in
main-interim branch and will fail there because it needs JGS code to
pass, but it is temporary until we change cron job into push/PR job for
main-js.

Signed-off-by: Gregory Shimansky <gregory.shimansky@intel.com>