[LoadStoreOpToLLVM] Transpose 2d load. (#4870)
To use the transpose 2d block io to load column major matrix from global
memory. (The column major matrix here could be generalized to the cases
that fast change dimension of register layout is not same as the fast
change dim on global memory.)

The 2d block io only can transpose the matrix of i32 type when load
matrix from memory to register. To transpose the matrix of type bits
number < 32, we need to further transpose matrix inside the register.

The steps the to load matrix with transposing with 2D block IO:
1. To load the matrix as d32 type matrix from memory with transposed to
register.
2. (Optional if scalar type < 32 bits) To transpose the `MxNxd32` to
`Mx(32/m)xNxdm` inside the register.

Right now we only use the bitcast operation to transpose the matrix of
which the width is equal to the threads per warp for step 2.

Further we will support more matrix of which the width is not equal to
the threads per warp.

Signed-off-by: Lu,Chengjun <chengjun.lu@intel.com>