[CRI] Support non-k-pack fp4 tt.dot_scaled ops to run on bdpas. (#835)
To unpack the fp4 which is not packed along k-dim by upcasting it to
fp16. It is safe because there is no rounding in upcast fp4 to fp16 and
the range is enlarged as well. The only drawback is that the BDPAS
operation is used four times compared to the original FP4 type.

Signed-off-by: Lu,Chengjun <chengjun.lu@intel.com>